feat: Include LiCl and CaCl2 in FluidProperties
This commit adds the LiCl and CaCl2 modules to the FluidProperties package. The modules are included in the Props.jl and Props_test.jl files. Additionally, the LiCl_test.jl and CaCl2_test.jl files are added to the test directory. These changes allow for the calculation of properties for aqueous solutions of LiCl and CaCl2, as described in the paper by Manuel R. Conde.