feat: Add calculate_T_sol function for LiCl and CaCl2
This commit adds the calculate_T_sol function for LiCl and CaCl2 in the FluidProperties package. The function allows for finding the temperature given the solute concentration and the activity coefficient. It is implemented in the CaCl2.jl and LiCl.jl files. The function uses the IntervalNonlinearProblem and solve functions from the DifferentialEquations.jl package to find the root of the equation. The temperature range for the root finding is set to 0.0°C to 95.0°C. This feature enables the calculation of the temperature barrier for aqueous solutions of LiCl and CaCl2, as described in the paper by Manuel R. Conde.